add arm

+++ include/ffmpeg/libavutil/common.h
@@ -204,8 +204,10 @@ static av_always_inline av_const int64_t av_clip64_c(int64_t a, int64_t amin, in
  */
 static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
 {
-    if (a&(~0xFF)) return (~a)>>31;
-    else           return a;
+    if (a&(~0xFF)) 
+      return (uint8_t)((~a)>>31);
+    else
+      return (uint8_t)(a);
 }
 
 /**
@@ -215,8 +217,10 @@ static av_always_inline av_const uint8_t av_clip_uint8_c(int a)
  */
 static av_always_inline av_const int8_t av_clip_int8_c(int a)
 {
-    if ((a+0x80U) & ~0xFF) return (a>>31) ^ 0x7F;
-    else                  return a;
+    if ((a+0x80U) & ~0xFF) return 
+      (int8_t)((a>>31) ^ 0x7F);
+    else
+      return (int8_t)(a);
 }
 
 /**
@@ -226,8 +230,10 @@ static av_always_inline av_const int8_t av_clip_int8_c(int a)
  */
 static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
 {
-    if (a&(~0xFFFF)) return (~a)>>31;
-    else             return a;
+    if (a&(~0xFFFF)) 
+      return (uint16_t)((~a)>>31);
+    else             
+      return (uint16_t)(a);
 }
 
 /**
@@ -235,10 +241,12 @@ static av_always_inline av_const uint16_t av_clip_uint16_c(int a)
  * @param a value to clip
  * @return clipped value
  */
-static av_always_inline av_const int16_t av_clip_int16_c(int a)
+static av_always_inline av_const int8_t av_clip_int16_c(int a)
 {
-    if ((a+0x8000U) & ~0xFFFF) return (a>>31) ^ 0x7FFF;
-    else                      return a;
+    if ((a+0x8000U) & ~0xFFFF) 
+      return (int8_t)((a>>31) ^ 0x7FFF);
+    else
+      return (int8_t)(a);
 }
 
 /**
